% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_project.R
\name{create_project}
\alias{create_project}
\title{Create a new project}
\usage{
create_project(path, ...)
}
\arguments{
\item{path}{Location of the new project.}

\item{...}{Not used currently.}
}
\description{
Create a new data processing or analysis project.
}
\details{
The file structure in this template was created by Emily Riederer:
https://emilyriederer.netlify.app/post/team-of-packages/
}
