% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_analysis.R
\name{new_analysis}
\alias{new_analysis}
\title{Create an .Rmd file in /analysis/}
\usage{
new_analysis(
  analysis_name,
  ...,
  analysis_path = here::here("analysis"),
  name_repair = TRUE
)
}
\arguments{
\item{analysis_name}{The name of the new analysis}

\item{...}{Currently unused}

\item{analysis_path}{Destination folder of the new .Rmd file}

\item{name_repair}{If TRUE, automatically make "analysis_name.Rmd" a simplified, valid file name}
}
\description{
Create an .Rmd file in /analysis/
}
