% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_src.R
\name{new_src}
\alias{new_src}
\title{Create an .R file for a new function in /src/}
\usage{
new_src(function_name, ..., src_path = here::here("src"), name_repair = FALSE)
}
\arguments{
\item{function_name}{The name of the new function}

\item{...}{Currently unused}

\item{src_path}{Destination folder of the new .R file}

\item{name_repair}{If TRUE, automatically make function_name valid}
}
\description{
Create an .R file for a new function in /src/
}
