% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/require_packages.R
\name{require_packages}
\alias{require_packages}
\title{Add packages to required_packages.txt}
\usage{
require_packages(..., .package_list = here::here("required_packages.txt"))
}
\arguments{
\item{...}{Names of packages to add to the list}

\item{.package_list}{File path of required_packages.txt}
}
\description{
Add packages to required_packages.txt
}
